Capture agent now be pushing its status to the capture admin status service
git-svn-id: https://opencast.jira.com/svn/MH/trunk@3852 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ modules/opencast-capture-service-api/trunk/src/main/java/org/opencastproject/capture/api/Scheduler.java
@@ -58,7 +58,7 @@ public interface Scheduler {
    * Gets the time between refreshes of the scheduling data
    * @return The time between refreshes of the scheduling data
    */
-  public int getPollingTime();
+  public long getPollingTime();
 
   /**
    * Enables polling for new calendar data
